requirements.txt
annotated-types   
anyio              
asgiref            
certifi            
charset-normalizer 
clickhouse-driver  
distro             
Django            
groq              
h11                
httpcore          
httpx            
idna               
pip                
psycopg2-binary    
pydantic           
pydantic_core      
python-decouple   
python-dotenv     
pytz              
requests          
sniffio           
sqlparse           
typing_extensions  
typing-inspection  
tzdata            
tzlocal            
urllib3            